Refactoring of modules
As funtionality was being added to Publication.py the code
became unmanageable. Therefore, code was refactor into 3
files. Code that is related to building_block were moved to
building_block.py and to factory_design_pattern.py and
mediator_design_pattern.py likewise.
