+++ modules/behavioural/__init__.py
@@ -1,0 +1,1 @@
+__all__ = ['mediator_design_pattern.py']
